# public/swagger.yaml
swagger: '2.0'

info:
  version: 1.0.0
  title: Bizhub
  description: Bizhub API resources and documentation

host: localhost:3000

basePath: /

schemes: 
  - http

consumes:
  - application/json

produces:
  - application/json

paths:
  /employers:
    get:
      description: |
        Gets list of `Employer` objects.
      operationId: getEmployers
      summary: get employers
      tags:
      - employer
      parameters:
      - $ref: "#/parameters/from"
      - $ref: "#/parameters/size"
      responses:
        200:
          description: Sucessful response
          schema:
            $ref: "#/definitions/Employers"
        500:
          $ref: "#/responses/Standard500ErrorResponse"
    post:
      description: |
        Create a new `Employer`
      operationId: createEmployer
      summary: create employer
      tags:
      - employer
      parameters:
      - name: employer
        in: body
        required: true
        description: employer to create
        schema:
          $ref: "#/definitions/NewEmployer"
      responses:
        201:
          description: Sucessfully created employer
          schema:
            $ref: "#/definitions/Employer"
  
  /employers/{employerId}:
    get:
      description: |
        Get a specific `Employer`
      operationId: getEmployerById
      summary: get employer by id
      tags:
      - employer
      parameters:
      - $ref: "#/parameters/employerId"
      responses:
        200:
          description: Sucessfull response
          schema:
            $ref: "#/definitions/Employer"
        500:
          $ref: "#/responses/Standard500ErrorResponse"
          
  /timesheets:
    get:
      description: |
        Gets list of `Timesheet` objects.
      operationId: getTimesheets
      summary: get timesheets
      tags:
      - timesheet
      parameters:
      - $ref: "#/parameters/from"
      - $ref: "#/parameters/size"
      responses:
        200:
          description: Successful response
          schema:
            $ref: "#/definitions/Timesheets"
        500:
          $ref: "#/responses/Standard500ErrorResponse"
                          
  /timesheets/{timesheetId}:
    get:
      description: |
          Gets a specific `Timesheet` object.
      operationId: getTimesheetById
      summary: 'get timesheet by id'
      tags:
      - timesheet
      parameters: 
      - $ref: "#/parameters/timesheetId"
      responses:
        200:
          description: Successful response
          schema:
            $ref: "#/definitions/Timesheet"
        500:
          $ref: "#/responses/Standard500ErrorResponse"
          
  /timesheets/{timesheetId}/entries:
    get:
      description: |
        Gets timesheet entries for a specific entry
      operationId: getTimesheetEntries
      summary: get timesheet entries
      tags:
      - timesheet
      - timesheet entry
      parameters:
      - $ref: "#/parameters/timesheetId"
      responses:
        200:
          description: Successfull response
          schema:
            $ref: "#/definitions/TimesheetEntries"
        500:
          $ref: "#/responses/Standard500ErrorResponse"
    post:
      description: |
        Create a Timesheet Entry for the specific timesheet
      operationId: createTimesheetEntry
      summary: create timesheet entry
      tags:
      - timesheet
      - timesheet entry
      parameters:
      - $ref: "#/parameters/timesheetId"
      - name: entry
        description: timesheet entry to create
        in: body
        required: true
        schema:
          $ref: "#/definitions/TimesheetEntry"
      responses:
        201:
          description: Successfully created timesheet entry
          schema:
            $ref: "#/definitions/TimesheetEntry"
        500:
          $ref: "#/responses/Standard500ErrorResponse"

definitions:
  Error:
    properties:
      code:
        type: string
      message:
        type: string
        
  NewEmployer:
    description: Employer post model
    required:
    - name
    - address_line_1
    - town_or_city
    - country
    - postcode
    properties:
      name:
        type: string
        description: name of the employer
        example: NHS
      address_line_1:
        type: string
        description: address line 1
        example: 2 Simple Street
      address_line_2:
        type: string
        description: address line 2
        example: Great Lever
      town_or_city:
        type: string
        description: town or city
        example: Bolton
      county:
        type: string
        example: Lancashire
      country:
        type: string
        example: United Kingdom
      postcode:
        type: string
        example: BL3 2JN
        
  Employer:
    description: An employer
    allOf:
    - $ref: '#/definitions/NewEmployer'
    required:
    - name
    - address_line_1
    - town_or_city
    - country
    - postcode
    properties:
      id:
        type: integer
        readOnly: true
        minimum: 0
        exclusiveMinimum: true
        example: 1
      url:
        type: string
        readOnly: true
        example: http://api.bizhub.io/employers/1234

  Employers:
    type: array
    items:
      $ref: "#/definitions/Employer"

  Timesheet:
    properties:
      id:
        type: integer
        readOnly: true
        minimum: 0
        exclusiveMinimum: true
        example: 1
      url:
        type: string
        readOnly: true
        example: http://api.bizhub.io/timesheets/1
      entries_url:
        type: string
        readOnly: true
        example: http://api.bizhub.io/timesheets/1/entries
      submit_date:
        type: string
        format: date-time
        description: date timesheet submitted
        
  Timesheets:
    type: array
    items:
      $ref: "#/definitions/Timesheet"
      
  TimesheetEntry:
    required:
    - start
    - end
    - break
    - rate_per_hour
    properties:
      id:
        type: integer
        readOnly: true
        minimum: 0
        exclusiveMinimum: true
        example: 1234
      url:
        type: string
        readOnly: true
        example: http://api.bizhub.io/timesheets/1/entries/1234
      start:
        type: string
        format: date-time
      end:
        type: string
        format: date-time
      break:
        type: number
        format: double
        example: 0.5
      rate_per_hour:
        type: number
        format: double
        example: 8.5
      employer:
        $ref: "#/definitions/Employer"
  
  TimesheetEntries:
    type: array
    items:
      $ref: "#/definitions/TimesheetEntry"

parameters:
  from:
    name: from
    in: query
    description: first element to return starting at 1
    required: false
    type: integer
    minimum: 0
    exclusiveMinimum: false
    default: 0

  size:
    name: size
    in: query
    description: max number of elements to return
    required: false
    type: integer
    minimum: 0
    exclusiveMinimum: true
    maximum: 100
    exclusiveMaximum: false
    default: 10

  timesheetId:
    name: timesheetId
    in: path
    description: ID of timesheet to return
    required: true
    type: integer
    format: int64
    
  employerId:
    name: employerId
    in: path
    description: ID of employer to return
    required: true
    type: integer
    format: int64

responses:
  Standard500ErrorResponse:
    description: An unexpected error occured.
    schema:
      $ref: "#/definitions/Error"